Add files via upload
This is the notebook which we used for the project. We compiled all
our code into one notebook that runs sequentially, although the file
size might be large and take time to load.

General Content:
* Data Handling, Cleaning
* Exploratory Data Analysis
* Collaborative Filtering (SVD, ALS, Cold Start)
* Content Filtering (TF-IDF, Cosine Similarity, Custom Defined Function)
* Timing and Location (Application to Business Case)
* Conclusion

Note: Code in Raw NBConvert take a long time to run. We have output to csv some of the results to save time. Those folders should be run.